# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: HASN Digital Services Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users
      parameters:
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
        - in: query
          name: search
          schema:
            type: string
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserList"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get a user by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update a user
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}/block:
    patch:
      operationId: blockUser
      tags: [users]
      summary: Block or unblock a user
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlockInput"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}/wallet:
    patch:
      operationId: adjustUserWallet
      tags: [users]
      summary: Adjust user wallet balance (admin only)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WalletAdjustInput"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  # Categories
...
  /wallet/transactions:
    get:
      operationId: listTransactions
      tags: [wallet]
      summary: List wallet transactions
      parameters:
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        "200":
          description: Transactions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionList"
  /wallet/recharge:
    post:
      operationId: requestRecharge
      tags: [wallet]
      summary: Request wallet recharge
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RechargeInput"
      responses:
        "201":
          description: Recharge request submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RechargeRequest"
  /wallet/recharge-requests:
    get:
      operationId: listRechargeRequests
      tags: [wallet]
      summary: List all recharge requests (admin)
      parameters:
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
      responses:
        "200":
          description: Requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RechargeRequestList"
  /wallet/recharge-requests/{id}/approve:
    patch:
      operationId: approveRecharge
      tags: [wallet]
      summary: Approve recharge request
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApproveRechargeInput"
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RechargeRequest"
  /wallet/recharge-requests/{id}/reject:
    patch:
      operationId: rejectRecharge
      tags: [wallet]
      summary: Reject recharge request
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RechargeRequest"

  # Orders
...
components:
  schemas:
...
    User:
      type: object
      required: [id, phone, name, role, isBlocked, walletBalance, createdAt]
      properties:
        id:
          type: integer
        phone:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [user, admin, super_admin]
        isBlocked:
          type: boolean
        walletBalance:
          type: number
        createdAt:
          type: string
    UserList:
      type: object
      required: [users, total, page, limit]
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    UserUpdate:
      type: object
      properties:
        name:
          type: string
        role:
          type: string
          enum: [user, admin, super_admin]
    BlockInput:
      type: object
      required: [isBlocked]
      properties:
        isBlocked:
          type: boolean

    # Category
...
    WalletAdjustInput:
      type: object
      required: [amount]
      properties:
        amount:
          type: number
        note:
          type: string
...
    Transaction:
      type: object
      required: [id, type, amount, status, createdAt]
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [recharge, purchase, refund]
        amount:
          type: number
        status:
          type: string
          enum: [pending, completed, failed, rejected]
        description:
          type: ["string", "null"]
        createdAt:
          type: string
    TransactionList:
      type: object
      required: [transactions, total, page, limit]
      properties:
        transactions:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    RechargeInput:
      type: object
      required: [amount, paymentMethodId]
      properties:
        amount:
          type: number
        paymentMethodId:
          type: integer
        transactionRef:
          type: string
        proofImageUrl:
          type: string
        notes:
          type: string
    RechargeRequest:
      type: object
      required: [id, userId, amount, status, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        userName:
          type: string
        userPhone:
          type: string
        amount:
          type: number
        approvedAmount:
          type: ["number", "null"]
        status:
          type: string
          enum: [pending, approved, rejected]
        transactionRef:
          type: ["string", "null"]
        proofImageUrl:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        paymentMethodId:
          type: integer
    RechargeRequestList:
      type: object
      required: [requests, total, page, limit]
      properties:
        requests:
          type: array
          items:
            $ref: "#/components/schemas/RechargeRequest"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    ApproveRechargeInput:
      type: object
      required: [approvedAmount]
      properties:
        approvedAmount:
          type: number

    # Orders
...